chore(core): token-diet phase 2-B — split mur-core/src/server.rs
Pure code movement, behavior unchanged. Splits the 2240-line
server.rs into a folder so every file satisfies the project's
800-line limit:

  server/
    mod.rs        (351)  shared types + envelope + build_router/run_server
                         + system handlers (ws / ui / health / rates)
    patterns.rs   (222)  /api/v1/patterns CRUD + filter
    workflows.rs  (299)  /api/v1/workflows CRUD + search + extract-from-session
    pipelines.rs  (215)  /api/v1/pipelines CRUD + run + validate
    sessions.rs   (220)  /api/v1/sessions list/get/events/delete/bulk/patch
    stats.rs      (138)  /api/v1/{stats,tags,links}
    search.rs      (61)  /api/v1/search + shared SearchRequest/SearchResult
    context.rs     (94)  /api/v1/{context,ingest,feedback}
    tests.rs      (748)  router integration tests

AppState helpers (pattern_store / workflow_store / pipeline_store) and
notify() are now `pub(super)` so resource modules can call them.
Handler request/response types referenced from build_router (e.g.
SessionInfo, BulkDeleteRequest, RunPipelineExprRequest) move to
`pub(super)` for the same reason. SearchRequest/SearchResult live in
search.rs and are imported by workflows::search_workflows.

cargo build -p mur-core + cargo test -p mur-core (1114 lib + 5 server
+ all integration suites) pass. cargo clippy -p mur-core --lib clean.
